Move trace pictures/icons into their own controllers
